requirements.txt
# Core Framework
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-JWT-Extended==4.5.3
Flask-Migrate==4.0.5
Flask-Cors==4.0.0
Flask-Limiter==3.5.0
Flask-Talisman==1.1.0
Gunicorn==23.0.0

# Database
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9

# Security & Utils
Werkzeug==3.0.1
python-dotenv==1.0.0
bleach==6.1.0
pokemontcgsdk==3.4.0

# Development & Testing
pylint==3.0.3
pytest==7.4.3